refactor: move time utilities to dedicated time.ts module
- Create src/lib/time.ts for time formatting utilities
- Move format_relative_time, format_utc_time, format_duration from utils.ts
- Rename tests/vitest/utils/time-formatting.test.ts → tests/vitest/time.test.ts
- Update imports to use new module location
- Export time module from main index.ts
- Better separation of concerns: utils for general utilities, time for time-specific
