+++ src/lib/index.ts
@@ -24,6 +24,7 @@ export * from './settings'
 export * from './structure'
 export * from './symmetry'
 export * from './theme'
+export * from './time'
 export { default as Trajectory } from './trajectory/Trajectory.svelte'
 export * from './utils'
 export * from './xrd'
